+++ apps/server/src/core/error/domain/domain-error-handler.ts
